♻️ [refactor] : 공용 인터페이스 및 함수 정리
