+++ .gitignore
@@ -6,4 +6,3 @@ dist
 test_case
 
 datas
-logs
