--- server/db/model/posts.ts
@@ -1,13 +1,0 @@
-import { RowDataPacket } from "mysql2";
-
-export interface IAdminPost {
-	id: number;
-	title: string;
-	author: string;
-	created_at: Date;
-	updated_at: Date;
-	isDelete: boolean;
-	is_private: boolean;
-}
-
-export interface IAdminPostRow extends IAdminPost, RowDataPacket {}
