--- shared/chats/message.ts
@@ -1,24 +1,0 @@
-import { IMessage } from "./dto";
-
-// 채팅방의 이전 메세지 기록 조회 ------------------------------------
-
-export interface IGetRoomMessageLogsRequest {
-	roomId: number; // 해당 방 번호
-}
-
-export interface IGetRoomMessageLogsResponse {
-	messageLogs: IMessage[]; // 채팅 로그
-}
-
-// 메세지 발신, 수신 -----------------------------------------------
-
-// 발신
-export interface ISendMessageRequest {
-	roomId: string;
-	message: string;
-}
-
-// 수신
-export interface IReceiveMessageResponse {
-	message: IMessage;
-}
