+++ shared/index.ts
@@ -1,16 +1,16 @@
-export * from "./posts/enums";
-export * from "./posts/mappers";
-export * from "./posts/models";
-export * from "./comments/mappers";
-export * from "./comments/models";
-export * from "./chats/chat-room";
-export * from "./chats/message";
-export * from "./chats/dto";
-export * from "./chats/mappers";
-export * from "../server/db/mapper/users_mapper";
-export * from "../server/db/mapper/posts_mapper";
-export * from "../server/db/mapper/logs_mapper";
-export * from "../server/db/mapper/stats_mapper";
-export * from "../server/db/model/logs";
-export * from "../server/db/model/stats";
-export * from "../server/db/model/users";
+export * from "./chat/chat_api";
+export * from "./chat/chat_mapper";
+export * from "./chat/chats";
+export * from "./community/admin/admin";
+export * from "./community/admin/admin_api";
+export * from "./community/admin/admin_mapper";
+export * from "./community/comments/comments";
+export * from "./community/comments/comment_mapper";
+export * from "./community/likes/likes";
+export * from "./community/likes/likes_mapper";
+export * from "./community/posts/posts";
+export * from "./community/posts/post_mapper";
+export * from "./community/users/users";
+export * from "./community/oauth/oauths";
+export * from "./community/oauth/oauth_mapper";
+export * from "./community/rbac";
