--- shared/posts/enums.ts
@@ -1,4 +1,0 @@
-export enum SortBy {
-	VIEWS = 1,
-	LIKES = 2,
-}
